Refs #126. Split up wordml module into openxml.packaging
