+++ tests/export/html/test_direct_formatting_bold_property.py
@@ -8,7 +8,7 @@ from __future__ import (
 
 from pydocx.test import DocumentGeneratorTestCase
 from pydocx.test.utils import WordprocessingDocumentFactory
-from pydocx.wordml import MainDocumentPart
+from pydocx.openxml.packaging import MainDocumentPart
 
 
 class DirectFormattingBoldPropertyTestCase(DocumentGeneratorTestCase):
