+++ tests/export/html/test_footnote.py
@@ -7,9 +7,9 @@ from __future__ import (
 )
 
 
+from pydocx.openxml.packaging import FootnotesPart, MainDocumentPart
 from pydocx.test import DocumentGeneratorTestCase
 from pydocx.test.utils import WordprocessingDocumentFactory
-from pydocx.wordml import FootnotesPart, MainDocumentPart
 
 
 class FootnoteTestCase(DocumentGeneratorTestCase):
