+++ tests/export/html/test_numbering.py
@@ -8,7 +8,7 @@ from __future__ import (
 
 from pydocx.test import DocumentGeneratorTestCase
 from pydocx.test.utils import WordprocessingDocumentFactory
-from pydocx.wordml import MainDocumentPart, NumberingDefinitionsPart
+from pydocx.openxml.packaging import MainDocumentPart, NumberingDefinitionsPart
 
 
 class NumberingTestCase(DocumentGeneratorTestCase):
