+++ tests/export/test_xml.py
@@ -9,7 +9,7 @@ import sys
 import time
 
 from pydocx.util.xml import find_all, parse_xml_from_string
-from pydocx.wordml import ImagePart
+from pydocx.openxml.packaging import ImagePart
 
 from pydocx.test import TranslationTestCase
 from pydocx.test.document_builder import DocxBuilder as DXB
