+++ tests/test_wordml.py
@@ -8,7 +8,7 @@ import unittest
 
 from pydocx.packaging import ZipPackage
 from pydocx.util.xml import xml_tag_split
-from pydocx.wordml import MainDocumentPart, WordprocessingDocument
+from pydocx.openxml.packaging import MainDocumentPart, WordprocessingDocument
 
 
 class ZipPackageTestCase(unittest.TestCase):
